@@ -1,33 +1,33 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
 	<PropertyGroup>
-		<TargetFramework>netcoreapp2.1</TargetFramework>
+		<TargetFramework>netcoreapp3.1</TargetFramework>
 		<DockerTargetOS>Linux</DockerTargetOS>
 		<GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
 		<AWSProjectType>Lambda</AWSProjectType>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Amazon.Lambda.AspNetCoreServer" Version="2.0.4" />
-		<PackageReference Include="Amazon.Lambda.Core" Version="1.0.0" />
-		<PackageReference Include="Amazon.Lambda.Serialization.Json" Version="1.3.0" />
-		<PackageReference Include="Antlr4.Runtime" Version="4.6.4" />
-		<PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.3.6" />
-		<PackageReference Include="AWSSDK.Kinesis" Version="3.3.6.1" />
-		<PackageReference Include="AWSSDK.Lambda" Version="3.3.16.2" />
-		<PackageReference Include="AWSSDK.S3" Version="3.3.20.2" />
-		<PackageReference Include="HtmlAgilityPack" Version="1.8.6" />
-		<PackageReference Include="Kitsune.Models" Version="1.0.6.1" />
-		<PackageReference Include="MessagePack" Version="1.7.3.4" />
-		<PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.0" />
-		<PackageReference Include="Microsoft.AspNetCore.ResponseCompression" Version="2.1.1" />
-		<PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.0" />
-		<PackageReference Include="MongoDB.Driver" Version="2.7.0" />
-		<PackageReference Include="murmurhash" Version="1.0.2" />
-		<PackageReference Include="StackExchange.Redis" Version="1.2.6" />
-		<PackageReference Include="System.Configuration.ConfigurationManager" Version="4.5.0" />
-		<PackageReference Include="System.Drawing.Common" Version="4.5.0" />
-		<PackageReference Include="System.Management" Version="4.5.0" />
+		<PackageReference Include="Amazon.Lambda.AspNetCoreServer" Version="6.0.3" />
+		<PackageReference Include="Amazon.Lambda.Core" Version="2.0.0" />
+		<PackageReference Include="Amazon.Lambda.Serialization.Json" Version="2.0.0" />
+		<PackageReference Include="Antlr4.Runtime" Version="4.6.6" />
+		<PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.7.1" />
+		<PackageReference Include="AWSSDK.Kinesis" Version="3.7.0.53" />
+		<PackageReference Include="AWSSDK.Lambda" Version="3.7.3.2" />
+		<PackageReference Include="AWSSDK.S3" Version="3.7.1.25" />
+		<PackageReference Include="HtmlAgilityPack" Version="1.11.36" />
+		<PackageReference Include="Kitsune.Models" Version="1.0.8" />
+		<PackageReference Include="MessagePack" Version="2.3.75" />
+<!--		<PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.0" />-->
+		<PackageReference Include="Microsoft.AspNetCore.ResponseCompression" Version="2.2.0" />
+		<PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.5" />
+		<PackageReference Include="MongoDB.Driver" Version="2.13.1" />
+		<PackageReference Include="murmurhash" Version="1.0.3" />
+		<PackageReference Include="StackExchange.Redis" Version="2.2.62" />
+		<PackageReference Include="System.Configuration.ConfigurationManager" Version="5.0.0" />
+		<PackageReference Include="System.Drawing.Common" Version="5.0.2" />
+		<PackageReference Include="System.Management" Version="5.0.0" />
 	</ItemGroup>
 
 	<ItemGroup>
